Move download links into separate component
